+++ tools/tapenade/run_gmm.cpp
@@ -1,43 +1,10 @@
 #include <stdlib.h>
 #include <iostream>
 #include "TapenadeGMM.h"
+#include "adbench/io.h"
 #include "adbench/shared/GMMData.h"
 #include "adbench/shared/utils.h"
 #include "json.hpp"
-
-void read_GMMInput_json(const char* fname, GMMInput &input) {
-  using json = nlohmann::json;
-  std::ifstream f(fname);
-  json data = json::parse(f);
-  input.d = data["d"].get<int>();
-  input.k = data["k"].get<int>();
-  input.n = data["n"].get<int>();
-  input.alphas = data["alpha"].get<std::vector<double>>();
-
-  auto means = data["means"].get<std::vector<std::vector<double>>>();
-  auto icf = data["icf"].get<std::vector<std::vector<double>>>();
-  auto x = data["x"].get<std::vector<std::vector<double>>>();
-  for (int i = 0; i < input.k; i++) {
-    input.means.insert(input.means.end(), means[i].begin(), means[i].end());
-    input.icf.insert(input.icf.end(), icf[i].begin(), icf[i].end());
-  }
-  for (int i = 0; i < input.n; i++) {
-    input.x.insert(input.x.end(), x[i].begin(), x[i].end());
-  }
-
-  input.wishart.gamma = data["gamma"].get<double>();
-  input.wishart.m = data["m"].get<int>();
-}
-
-void write_GMMOutput_F_json(std::ostream& f, GMMOutput &output) {
-  using json = nlohmann::json;
-  f << json(output.objective);
-}
-
-void write_GMMOutput_J_json(std::ostream& f, GMMOutput &output) {
-  using json = nlohmann::json;
-  f << json(output.gradient);
-}
 
 int main(int argc, char* argv[]) {
   if (argc != 3 ||
@@ -66,14 +33,14 @@ int main(int argc, char* argv[]) {
     clock_gettime( CLOCK_REALTIME, &finish );
 
     GMMOutput output = gmm.output();
-    write_GMMOutput_F_json(std::cout, output);
+    write_GMMOutput_objective_json(std::cout, output);
   } else {
     clock_gettime( CLOCK_REALTIME, &start );
     gmm.calculate_jacobian(1);
     clock_gettime( CLOCK_REALTIME, &finish );
 
     GMMOutput output = gmm.output();
-    write_GMMOutput_J_json(std::cout, output);
+    write_GMMOutput_jacobian_json(std::cout, output);
   }
   std::cout << std::endl;
